requirements.txt
python-gitlab>=3.9
flask
typer
rich